Option to enable -PS status feature but start in logging mode (#1701)
* Option to enable -PS status feature but start in logging

Currently if you don't use the -ps or --print-status command line
option, there is no way to get into the status screen later if desired.
The is because the thread to enable keyboard monitoring is only
initiated if the -ps option is used. This change adds a 'logs' options
to the -ps and --prints-status command line option to enable the -PS
feature yet start in logging mode. It launches the keyboard monitoring
thread but defaults to logging mode.

* Fix travis