Quick fixes related to captchas (#1823)
* Bookmarklet minor fixes

Bookmarklet was using browser's URL to generate the bookmarklet script instead of using --manual-captcha-domain. This shouldn't be a problem for most people but in uncommon configurations it may result in some confusion. Now we will use -mcd to generate the bookmarklet.
Removed duplicate paragraph in "Handling Captchas" documentation.

* Webhook message quick fix.

Documentation minor text fix.