Ignore specific pokemon (do not store to DB) (#1962)
* Implemented an ignore list.  This will remove pokemon them from
the list returned in parse_map and then the rest of code never
sees them

* Removed a line of debug that slipped through.

* Removed the temporary array and just iterating through
cell.get('wild_pokemons',[]) directly.

* Had to change things around.  Through extensive tests, it was
discovered that by completely ignoring them in the original list
that it negatively impacted spawn rate detection.  This is bad news
after a -cd.  So, instead, I put a simple check farther down. this
will still prevent the pokemon from being stored in the database
and it will also not be sent to webhooks.  In my testing, on a -cd
this resolved the issue and all spawn points were found, even with
ignored pokemon.

* Fixed whitespace issue for Travis.

* Updated (and reformatted) command line documentation

* Change the whitelist to read only from a file. This is consistent
with the (new) encounter whitelist. Also, noticed redudant assignments
of pokemon_id in parse_map, so I cleaned those up.
Updated documentation.

* Tweaked documentation

* Removed the mutually exclusive group, anmd checked for the ignore list
argument before checking for ignored pokemon.  Obviously, this was
going to throw exceptions.

* Fixed extraneous commands in documentation

* Updated documentation, and command output clarification.

* Added an empty set for the ignore list.

* Travis...

* Whitespace cleanup

* Added filtering counter, and information

* Whitespace issue

* Fixed the log output.

* Updated docs after rebase

* added 's' to raid, and replaced tabs with spaces in commandline output
file.

* Ooops.  Still had an old p[] reference instead of p.

* Updated per Seb's requested changes.

* Fix descriptions.

* Redundant is redundant.

* Consistency fix.